Solve exercise 8 and 9
Extracted the desired 7 columns and stored the results of one page in a dataframe for ex. 8.
In exercise 9 I created a function and then looped over all pages. After each iteration I extracted the 7 desired columns and stored all results in a data frame. Did some adjustments of the dataframe, too. For example longitude and latitude need to be type numeric for further usage.